Make changes suggested in a recent code review
The changes include replacing preprocessor defines
TIME_BEFORE_WINDOW_CLOSES_AFTER_DOWNLOADS and FIXED_WIDGET_HEIGHT
with static const ints of the same names, replacing the macro
FIND_SENDER_INDEX with the function findSenderIndex() and changing
a few "new statements" so that the created objects have a pointer
to their parent object. This assures the objects will be destroyed
when their parent object is destroyed.