Revert previous commit, validate url input instead
We cannot use http:// scheme because this causes the
server to return a 301 redirect what tmcclient does not support.

Instead, we ask the user to use https:// in the login screen. For
testing, the settings loaded from disk are not validated, so testers
can edit the configuration file to include http:// if need be.