Write qsettings (#47)
* Usability fixes in Settings/exercise management

Made organization list more beautifuler and added
helpful information messages to user

* Small changes
- Empty password field
- Default checked status on download list
- Change order of login/settings in menu

* Disable download/submit buttons when not logged in

* Save exercise data to qsettings
- Do not load it on launch if the directory has been removed

* Fixed a bug where exercise name wasn't being loaded properly

* Enable automatic updates by setting the interval

* Validate url input in LoginWidget

We cannot use http:// scheme because this causes the
server to return a 301 redirect what tmcclient does not support.

Instead, we ask the user to use https:// in the login screen. For
testing, the settings loaded from disk are not validated, so testers
can edit the configuration file to include http:// if need be.